fix: result에서 protein hits description이 올바르게 뜨지 않던거 수정 + peptide hits 정렬을 컨트롤러로 옮김
